Refactor with DigitalCollectionsMetadata class
